fix: 🐞 fixing component
