+++ packages/backend/.gitignore
@@ -11,7 +11,6 @@ server_generate/bucket_files
 
 /.pnp
 .pnp.js
-*.jar
 *.env
 *.log
 *.p8
