+++ packages/frontend/src/components/ui/admin/AuthorsUsers/AuthorsUsers.view.jsx
@@ -3,7 +3,7 @@ import { useStyles } from '../../../../styles/MainStyles';
 import { Grid, Typography } from '@material-ui/core';
 import TableComponent from '../../../Reusables/TableComponent';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import UserInfo from '../../../AdminScreens/UserInfo';
+import UserInfo from '../UserInfo/UserInfo';
 
 const AuthorsUsersView = (props) => {
 	const classes = useStyles();
