+++ packages/frontend/src/components/ui/admin/UserInfoBlock/UserInfoBlock.js
@@ -1,0 +1,1 @@
+
